pyproject.toml
[project]
name = "recipe-formatter"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Kyle Kestell", email = "kyle@kestell.org"},
]
dependencies = [
    "openai>=1.36.0",
    "instructor>=1.3.5",
    "python-slugify>=8.0.4",
    "python-recipy>=0.1.0",
    "llama-cpp-python>=0.3.1",
]
requires-python = ">=3.9, <=3.13"
readme = "README.md"
license = {text = "0BSD"}


[tool.pdm]
distribution = false

[tool.pdm.dev-dependencies]
dev = [
    "nuitka>=2.3.11",
]

[tool.pdm.scripts]
start = "python -m src.recipe_formatter.main"
